Update README.md
This project consists of a book search system. In the navigation bar there is an input field where the user types the title of the work and the system searches and displays in real time the list of books with the keyword, each book has the cover image, the description and when clicking on them a A new browser tab opens and this book's Google store page loads.